+++ apps/api/v2/src/ee/provider/outputs/verify-client.output.ts
@@ -1,10 +1,20 @@
 import { ApiProperty } from "@nestjs/swagger";
-import { IsEnum } from "class-validator";
+import { Type } from "class-transformer";
+import { IsEnum, IsNotEmptyObject, ValidateNested } from "class-validator";
 
 import { SUCCESS_STATUS, ERROR_STATUS } from "@calcom/platform-constants";
+import { PlatformOAuthClientDto } from "@calcom/platform-types";
 
 export class ProviderVerifyClientOutput {
   @ApiProperty({ example: SUCCESS_STATUS, enum: [SUCCESS_STATUS, ERROR_STATUS] })
   @IsEnum([SUCCESS_STATUS, ERROR_STATUS])
   status!: typeof SUCCESS_STATUS | typeof ERROR_STATUS;
+
+  @ApiProperty({
+    type: PlatformOAuthClientDto,
+  })
+  @IsNotEmptyObject()
+  @ValidateNested()
+  @Type(() => PlatformOAuthClientDto)
+  data!: PlatformOAuthClientDto;
 }
