+++ apps/api/v2/src/ee/provider/provider.controller.ts
@@ -30,16 +30,18 @@ export class CalProviderController {
 
   @Get("/:clientId")
   @HttpCode(HttpStatus.OK)
-  async verifyClientId(@Param("clientId") clientId: string): Promise<ProviderVerifyClientOutput> {
+  async getClient(@Param("clientId") clientId: string): Promise<ProviderVerifyClientOutput> {
     if (!clientId) {
       throw new NotFoundException();
     }
+
     const oAuthClient = await this.oauthClientRepository.getOAuthClient(clientId);
 
     if (!oAuthClient) throw new UnauthorizedException();
 
     return {
       status: SUCCESS_STATUS,
+      data: oAuthClient,
     };
   }
