+++ packages/platform/atoms/hooks/useOAuthClient.ts
@@ -2,7 +2,7 @@ import type { AxiosError } from "axios";
 import { useState, useEffect } from "react";
 import { usePrevious } from "react-use";
 
-import type { ApiResponse } from "@calcom/platform-types";
+import type { ApiResponse, PlatformOAuthClientDto } from "@calcom/platform-types";
 
 import http from "../lib/http";
 
@@ -16,6 +16,7 @@ export interface useOAuthClientProps {
 export const useOAuthClient = ({ clientId, apiUrl, refreshUrl, onError, onSuccess }: useOAuthClientProps) => {
   const prevClientId = usePrevious(clientId);
   const [isInit, setIsInit] = useState<boolean>(false);
+  const [orgId, setOrgId] = useState<number>(0);
 
   useEffect(() => {
     if (apiUrl && http.getUrl() !== apiUrl) {
@@ -31,10 +32,13 @@ export const useOAuthClient = ({ clientId, apiUrl, refreshUrl, onError, onSucces
     if (clientId && http.getUrl() && prevClientId !== clientId) {
       try {
         http
-          .get<ApiResponse>(`/provider/${clientId}`)
-          .then(() => {
+          .get<ApiResponse<PlatformOAuthClientDto>>(`/provider/${clientId}`)
+          .then((response) => {
             onSuccess();
             http.setClientIdHeader(clientId);
+            if (response.data.status === "success") {
+              setOrgId(response.data.data.organizationId);
+            }
           })
           .catch((err: AxiosError) => {
             if (err.response?.status === 401) {
@@ -47,5 +51,5 @@ export const useOAuthClient = ({ clientId, apiUrl, refreshUrl, onError, onSucces
     }
   }, [clientId, onError, prevClientId, onSuccess]);
 
-  return { isInit };
+  return { isInit, orgId };
 };
