+++ packages/platform/types/oauth-clients.ts
@@ -1,4 +1,5 @@
-import { IsArray, IsNumber, IsOptional, IsBoolean, IsString } from "class-validator";
+import { ApiProperty } from "@nestjs/swagger";
+import { IsArray, IsNumber, IsDate, IsOptional, IsBoolean, IsString } from "class-validator";
 import { z } from "zod";
 
 export class CreateOAuthClientInput {
@@ -49,3 +50,39 @@ export const userSchemaResponse = z.object({
 });
 
 export type UserResponse = z.infer<typeof userSchemaResponse>;
+
+export class PlatformOAuthClientDto {
+  @ApiProperty({ example: "clsx38nbl0001vkhlwin9fmt0" })
+  @IsString()
+  id!: string;
+
+  @ApiProperty({ example: "MyClient" })
+  @IsString()
+  name!: string;
+
+  @ApiProperty({ example: "secretValue" })
+  @IsString()
+  secret!: string;
+
+  @ApiProperty({ example: 3 })
+  @IsNumber()
+  permissions!: number;
+
+  @ApiProperty({ example: "https://example.com/logo.png", required: false })
+  @IsOptional()
+  @IsString()
+  logo!: string | null;
+
+  @ApiProperty({ example: ["https://example.com/callback"] })
+  @IsArray()
+  @IsString({ each: true })
+  redirectUris!: string[];
+
+  @ApiProperty({ example: 1 })
+  @IsNumber()
+  organizationId!: number;
+
+  @ApiProperty({ example: "2024-03-23T08:33:21.851Z", type: Date })
+  @IsDate()
+  createdAt!: Date;
+}
